consumer/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
confluent-kafka==2.13.0
dnspython==2.8.0
fastapi==0.129.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
redis==7.2.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0